[reconfigurator] Planner/builder cleanup 2/n: extract DisksEditor
